Clean up app.js by creating component separately
