+++ src/main/imageuploadui/src/App.js
@@ -1,21 +1,7 @@
 import React, {useState, useEffect} from "react";
 import './App.css';
 import axios from 'axios'
-
-const UserProfiles = () => {
-  const fetchUserProfiles =() => {
-    axios.get("http://localhost:8080/api/v1/user-profile")
-    .then(resp => {
-      console.log(resp);
-    })
-  }
-
-  useEffect(() => {
-    fetchUserProfiles();
-  }, [])
-
-  return <h1>Hello Image</h1>
-}
+import UserProfiles from './Component/UserProfiles';
 
 function App() {
   return (
